Update BFS
Breadth First Search (BFS) is an algorithm for traversing or searching tree or graph data
structures. It explores all the nodes at the present depth before moving on to the nodes at the next
depth level. Breadth-first search is a graph traversal algorithm that starts traversing the graph
from the root node and explores all the neighboring nodes. Then, it selects the nearest node and
explores all the unexplored nodes. While using BFS for traversal, any node in the graph can be
considered as the root node.